+++ kparser2.Analytics/Models.fs
@@ -87,6 +87,7 @@ type Interaction =
       Value: int
       Success: string
       CommandNo: int
+      SpellId: int option
       MessageId: int
       IsProc: bool
       ProcValue: int